move remaining cli collections code to AC
