+++ apps/cli/src/admin-console/models/response/collection.response.ts
@@ -1,7 +1,7 @@
 import { CollectionView } from "@bitwarden/admin-console/common";
 import { CollectionWithIdExport } from "@bitwarden/common/models/export/collection-with-id.export";
 
-import { BaseResponse } from "../../models/response/base.response";
+import { BaseResponse } from "../../../models/response/base.response";
 
 export class CollectionResponse extends CollectionWithIdExport implements BaseResponse {
   object: string;
