+++ apps/cli/src/commands/get.command.ts
@@ -37,6 +37,7 @@ import { FolderView } from "@bitwarden/common/vault/models/view/folder.view";
 import { KeyService } from "@bitwarden/key-management";
 
 import { OrganizationCollectionRequest } from "../admin-console/models/request/organization-collection.request";
+import { CollectionResponse } from "../admin-console/models/response/collection.response";
 import { OrganizationCollectionResponse } from "../admin-console/models/response/organization-collection.response";
 import { OrganizationResponse } from "../admin-console/models/response/organization.response";
 import { SelectionReadOnly } from "../admin-console/models/selection-read-only";
@@ -46,7 +47,6 @@ import { TemplateResponse } from "../models/response/template.response";
 import { SendResponse } from "../tools/send/models/send.response";
 import { CliUtils } from "../utils";
 import { CipherResponse } from "../vault/models/cipher.response";
-import { CollectionResponse } from "../vault/models/collection.response";
 import { FolderResponse } from "../vault/models/folder.response";
 
 import { DownloadCommand } from "./download.command";
